\name{cluster.summary}
\alias{cluster.summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
##  ##function to do ... ##
}
\description{
##  ## A concise (1-5 lines) description of what the function does. ##
}
\usage{
cluster.summary(e.cutoff = 0.01, nrow.cutoff = 5, seq.type = "upstream", plot = F, sort = c("score.norm", "score", "resid", "e.value1", "e.value2", "nrow"), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{e.cutoff}{
##     ##Describe \code{e.cutoff} here##
}
  \item{nrow.cutoff}{
##     ##Describe \code{nrow.cutoff} here##
}
  \item{seq.type}{
##     ##Describe \code{seq.type} here##
}
  \item{plot}{
##     ##Describe \code{plot} here##
}
  \item{sort}{
##     ##Describe \code{sort} here##
}
  \item{\dots}{
##     ##Describe \code{\dots} here##
}
}
\details{
##  ## If necessary, more details than the description above ##
}
\value{
##  ~Describe the value returned
##  If it is a LIST, use
##  \item{comp1 }{Description of 'comp1'}
##  \item{comp2 }{Description of 'comp2'}
## ...
}
\references{
## ~put references to the literature/web site here ~
}
\author{
##  ##who you are##
}
\note{
##  ##further notes##
}


\seealso{
## ##objects to See Also as \code{\link{help}}, ##~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
