\name{plotClust}
\alias{plotClust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
##  ##function to do ... ##
}
\description{
##  ## A concise (1-5 lines) description of what the function does. ##
}
\usage{
plotClust(k, w.motifs = T, all.conds = F, title = NULL, o.genes = NULL, dont.plot = F, network = "all", short.names = organism == "sce", seq.type = names(mot.weights), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{k}{
##     ##Describe \code{k} here##
}
  \item{w.motifs}{
##     ##Describe \code{w.motifs} here##
}
  \item{all.conds}{
##     ##Describe \code{all.conds} here##
}
  \item{title}{
##     ##Describe \code{title} here##
}
  \item{o.genes}{
##     ##Describe \code{o.genes} here##
}
  \item{dont.plot}{
##     ##Describe \code{dont.plot} here##
}
  \item{network}{
##     ##Describe \code{network} here##
}
  \item{short.names}{
##     ##Describe \code{short.names} here##
}
  \item{seq.type}{
##     ##Describe \code{seq.type} here##
}
  \item{\dots}{
##     ##Describe \code{\dots} here##
}
}
\details{
##  ## If necessary, more details than the description above ##
}
\value{
##  ~Describe the value returned
##  If it is a LIST, use
##  \item{comp1 }{Description of 'comp1'}
##  \item{comp2 }{Description of 'comp2'}
## ...
}
\references{
## ~put references to the literature/web site here ~
}
\author{
##  ##who you are##
}
\note{
##  ##further notes##
}


\seealso{
## ##objects to See Also as \code{\link{help}}, ##~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
