\name{save.cmonkey.env}
\alias{save.cmonkey.env}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
##  ##function to do ... ##
}
\description{
##  ## A concise (1-5 lines) description of what the function does. ##
}
\usage{
save.cmonkey.env(env = NULL, file = NULL, restore = T, verbose = T)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{env}{
##     ##Describe \code{env} here##
}
  \item{file}{
##     ##Describe \code{file} here##
}
  \item{restore}{
##     ##Describe \code{restore} here##
}
  \item{verbose}{
##     ##Describe \code{verbose} here##
}
}
\details{
##  ## If necessary, more details than the description above ##
}
\value{
##  ~Describe the value returned
##  If it is a LIST, use
##  \item{comp1 }{Description of 'comp1'}
##  \item{comp2 }{Description of 'comp2'}
## ...
}
\references{
## ~put references to the literature/web site here ~
}
\author{
##  ##who you are##
}
\note{
##  ##further notes##
}


\seealso{
## ##objects to See Also as \code{\link{help}}, ##~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
