\name{cMonkey-package}
\alias{cMonkey-package}
\alias{cMonkey}
\docType{package}
\title{
cMonkey integrated biclustering algorithm
}
\description{
cMonkey integrated biclustering algorithm for learning co-regulated gene modules
}
\details{
\tabular{ll}{
Package: \tab cMonkey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-10\cr
License: \tab GPL-2\cr
}
\url{https://github.com/dreiss-isb/cmonkey}
}
\author{
David J Reiss, Institute for Systems Biology

Maintainer: \email{dreiss@systemsbiology.org}
}
\references{
"Integrated biclustering of heterogeneous genome-wide datasets for the inference of global regulatory networks",
by David J Reiss, Nitin S Baliga and Richard Bonneau: \url{http://www.biomedcentral.com/1471-2105/7/280} 
}
\keyword{ package }
\seealso{
\url{https://github.com/dreiss-isb/cmonkey}
}
\examples{
 
}
